Separate models into apps
