+++ backend/mypage/views.py
@@ -3,12 +3,13 @@ from django.views.decorators.http import require_http_methods
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django.http.response import JsonResponse
-from main.models import ArticleGroupScrap, ArticleMemberScrap, MemberComment, GroupComment, MyIdolMember, MyIdolGroup
+from .models import ArticleGroupScrap, ArticleMemberScrap, MyIdolMember, MyIdolGroup
+from search_result.models import MemberComment, GroupComment
 
 LOGIN_PATH = "/"
 TYPE = "type"
-TYPE_MEMBER = 'member'
-TYPE_GROUP = 'group'
+TYPE_MEMBER = "member"
+TYPE_GROUP = "group"
 
 
 @login_required(login_url=LOGIN_PATH)
@@ -18,13 +19,13 @@ def myCmtGet(request):
     grpCmt = [GroupComment.objects.filter(user=request.user).values()]
     for cmt in mmbrCmt:
         cmt[TYPE] = TYPE_MEMBER
-        
+
     for cmt in grpCmt:
         cmt[TYPE] = TYPE_GROUP
-        
+
     myComment = mmbrCmt + grpCmt
     return JsonResponse(myComment, safe=False)
-    
+
 
 @login_required(login_url=LOGIN_PATH)
 @require_http_methods(["GET"])
@@ -33,25 +34,25 @@ def myIdolGet(request):
     myGrp = [MyIdolGroup.objects.filter(user=request.user)]
     for myIdol in myMmbr:
         myIdol[TYPE] = TYPE_MEMBER
-        
+
     for myIdol in myGrp:
         myIdol[TYPE] = TYPE_GROUP
-        
+
     myIdol = myMmbr + myGrp
     return JsonResponse(myIdol, safe=False)
-    
-    
-@login_required(login_url=LOGIN_PATH)    
+
+
+@login_required(login_url=LOGIN_PATH)
 @require_http_methods(["GET"])
 def myArtclGet(request):
     mmbrArtcl = [ArticleMemberScrap.objects.filter(user=request.user)]
     grpArtcl = [ArticleGroupScrap.objects.filter(user=request.user)]
     for artcle in mmbrArtcl:
         artcle[TYPE] = TYPE_MEMBER
-        
+
     for artcle in grpArtcl:
         artcle[TYPE] = TYPE_GROUP
-        
+
     myArticle = mmbrArtcl + grpArtcl
     return JsonResponse(myArticle, safe=False)
 
@@ -63,12 +64,14 @@ def mmbrArtcleDelete(request, article_id):
     myMmbrArtcl.delete()
     return HttpResponse(status=200)
 
+
 @login_required(login_url=LOGIN_PATH)
 @require_http_methods(["DELETE"])
 def grpArtcleDelete(request, article_id):
     myGrpArtcl = get_object_or_404(ArticleGroupScrap, pk=article_id)
     myGrpArtcl.delete()
     return HttpResponse(status=200)
+
 
 @login_required(login_url=LOGIN_PATH)
 @require_http_methods(["DELETE"])
@@ -77,10 +80,13 @@ def mmbrIdolDelete(request, idol_id):
     myMmbrIdol.delete()
     return HttpResponse(status=200)
 
+
 @login_required(login_url=LOGIN_PATH)
 @require_http_methods(["DELETE"])
 def grpIdolDelete(request, idol_id):
     myGrpIdol = get_object_or_404(MyIdolGroup, pk=idol_id)
     myGrpIdol.delete()
     return HttpResponse(status=200)
+
+
 # Create your views here.
