+++ backend/video/models.py
@@ -1,3 +1,5 @@
 from django.db import models
 
-# Create your models here.
+
+class VideoRecognitionModel(models.Model):
+    pass
